refactor(ds): update components and story
