+++ libs/design-system-components/src/components/AppCreationForm/index.tsx
@@ -131,7 +131,6 @@ const AppCreationForm: React.FC<AppCreationFormProps> = ({
                 setSelected={onChange}
                 ref={ref}
                 required={true}
-                requiredFieldAsteriskColor={{ light: 'errorLight', dark: 'errorDark' }}
               />
             )}
             defaultValue={extensionTypes[0]}
@@ -153,7 +152,6 @@ const AppCreationForm: React.FC<AppCreationFormProps> = ({
                 onChange={onChange}
                 inputRef={ref}
                 required={true}
-                requiredFieldAsteriskColor={{ light: 'errorLight', dark: 'errorDark' }}
               />
             )}
             defaultValue=""
@@ -175,7 +173,6 @@ const AppCreationForm: React.FC<AppCreationFormProps> = ({
                 onChange={onChange}
                 inputRef={ref}
                 required={true}
-                requiredFieldAsteriskColor={{ light: 'errorLight', dark: 'errorDark' }}
               />
             )}
             defaultValue=""
@@ -194,7 +191,6 @@ const AppCreationForm: React.FC<AppCreationFormProps> = ({
                   setSelected={onChange}
                   ref={ref}
                   required={true}
-                  requiredFieldAsteriskColor={{ light: 'errorLight', dark: 'errorDark' }}
                 />
               </>
             )}
@@ -216,7 +212,6 @@ const AppCreationForm: React.FC<AppCreationFormProps> = ({
                   onChange={onChange}
                   inputRef={ref}
                   required={true}
-                  requiredFieldAsteriskColor={{ light: 'errorLight', dark: 'errorDark' }}
                 />
               )}
               defaultValue=""
